backend/requirements.txt
Flask
ultralytics
torch
torchvision
opencv-python
numpy
pandas
Pillow
gdown
requests
tqdm
PyYAML
scipy
gunicorn
python-dotenv